Add linting
The back-end is going to be linted by rubocop and the front end by
eslint with formatting help from prettier.  The Javascript will be
formatted to the standard format, though I've elected to remove the
space between the function name and parentheses because I couldn't
figure out how to make prettier enforce that.