Implement back-end tracking of metadata and data
The metadata, controlled by the Metadata component in the corner, now
has a closed loop with the back end, so you can name your polls and have
other people see it.  The UI for this kind of sucks, but I'll fix it
later if I have time.

There's also tracking of the calendar data on the back-end, but the
front-end for visualizing the poll results will be its own change.